requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0

# GraphQL
strawberry-graphql[fastapi]==0.217.0

# Database
sqlalchemy[asyncio]==2.0.25
aiomysql==0.2.0
alembic==1.13.1

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Security
slowapi==0.1.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging
structlog==24.1.0

# Monitoring
prometheus-fastapi-instrumentator==6.1.0

# Utils
python-dotenv==1.0.0
greenlet==3.0.3
